requirements.txt
pandas
numpy
scikit-learn
matplotlib
joblib
torch
rapidfuzz
openai
python-dotenv